Add ability to capture raw messages
The listener will now capture raw messages and write them to
files which we assume a collection agent will gather and
send away to be saved.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I172c34976cf59853f0fb0dd7e0d7912197fdb0de

Add copy to support xdev directories

If the rename of a raw capture file fails with because
src and target are not on the same filesystem, then we
will copy the file.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I378dae46011c5480aecde62d484c98377f39026a