Add CI to build both images
This change adds the needed directory structure and config
files under the ci directory to build both the mc-core and
the mc_listener images.  Changes are needed in the ci
management repo to "activate" these, so this is part 1
which sets the stage for the multi-image build.

Also added a brief README at the top level.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ief76d4e1a8811d06464af0d5d158de27bf5b0221

Make core directory name under ci consistent with root source dir

Change-Id: I797c0f6373dd12669a032e619a64dafc1322a401

Add correct tag for mc-core

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I0087ced6c876c3b91ca74f1b6e13a909ab765d4d

Add top level readme

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I0c1eb11d26ab86f0bda1a84256587bc2da460d53