Slight realignment of CI files
To better work with the CI environment, the Dockerfiles and the
container-tag.yaml files will now need to live in the top of
each source directory (e.g. mc-core and sidecars/listener). This
change adds the correct Dockerfile for the listener, and the
tag file for both (the Dockerfile in mc-core was already there
and correct).

The following this change the contents of the ci directory are
likely uneeded, and a separte change will be introduced to
remove any/all which are not necessary.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ie508b1cd4629dca4629a07d48903dc0f6048e262