Update patch set 1
Patch Set 1:

(2 comments)

Also, good catch on the base for the mc-core docker file.  I caught that when I put the dockerfile in ci, a copy from the base directory, but in scrambling this morning to get everything under the top of each source dir I forgot to compare and thus didn't notice. 

I'll change that so we can test, then I can cobble together the listener changes to use the 'official' RMR versions.

Patch-set: 1