Remove environment setup from listener dockerfile
The dockerfile for the lisener image now uses a project
image which includes the necessary NNG and RMR libraries
and thus does not need to build them.  The invocation of
the build script to create these libraries was not remvoed
from the docker file when the project image was added.

This change removes the copy and execution of the script
from the dockerfile.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I15735813d37debc014657a358aa2b2fad319da49