Add FIFO pre-create to the replay startup script
The replay run script now provides the ability to pre-create
the FIFOs either with an explicit list, or by reading the
input file to suss out the message types it contains.

The run script also provides a way to gate and/or delay
the start of the replay binary in order to allow for better
synchronisation with a FIFO reading process.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I0c47ff4724a2877614cfef3e6d68033e1146b4a6