Update patch set 1
Patch Set 1:

(2 comments)

recheck

From the log it appears that it's not able to find ric-app-mc:1.0.0 in the repo:
nexus3.o-ran-sc.org:10004/o-ran-sc/ric-app-mc:1.0.0.    I know that I can run docker pull on the mc-listener that is there (docker pull nexus3.o-ran-sc.org:10004/o-ran-sc/ric-app-mc-listener:1.3.0) and it finds it, so the question is whether or not there was ever an image built for ric-app-mc.   


I edited the change to test by just having it do the listener image which I know is there; the recheck should cause it to rerun the jenkins things wth the change

Patch-set: 1
Label: Verified=0