Bundle the listener into the core image
This change causes the mc_listener application to be included
in the core MC image.  This is accomplished by basing the
final image on the listener container and by adding a "container
start" script which will start the "sidecars" (listener in this
case) and finally starts the core application.

The container version has been bumped to 2.0.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I983cacbaef4cfd95144664b05043b39e478b7827