Add CMake to drive unit tests in sidecars
In order to support sonar vetting, the unit tests must be driven by
CMake with a CMake file at the repo root. This change adds a CMake
file at the top level which will initially drive the unit tests
in the sidecars. This can be extended to include the core application.

Issue-ID: RIC-632

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I5707fa05bcd642a4f18473f05b4467a533b3cf34