Cleanup listener based on sonar grumblings
This change attempts to address the bugs and smells in the listener.
A fair few of the "bugs" were unit test coverage related (passing a
nil pointer) and to avoid those the test code was moved into a new
test directory, but it might need to be overtly excluded inthe
CI setup.  The other bugs were addressed.

Issue-ID: RIC-632

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ifbc9234e4408db3cb7111dc4a4ad9e9e268341c9