Add publish script to lister test set
The coverage files need to have complete paths (relative to the
repo root) in order for Sonar to match them with the analysis information
that it captured during build. This change adds a publish script to
the listener test scripts that will push the coverage files into
the reports directory and will fix up the internal source module name
for Sonar to match.

Issue-ID: RIC-632

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I438844a176b34f6026e3fcf2ad084ea303a05cbb