Add new xapp manager interface
The new interface for xAPP manager requires that the xAPP send
a "regestration" when the application/container starts, and an
unregister when shutting down.  This change adds a new "xam"
(xappManager) script which is invoked by the container start
script to do this.  There are also a couple of other utilities
which make bundling the confi data for the curl call easier.

Issue-ID: RIC-739

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I5eb4173e6bce7e7aca39d7afac36a9fc949d9a47