Correct bug in listener not detecting eagain/eintr correctly
It was possible for the listener to write a partial payload into
the pipe and completely bugger the reader with an out of synch
state. This change adds the necessary code for the listener to
properly handle an interrupted write operation while still
detecting a 'no reader' sate on a pipe.

Issue-ID: RIC-6768

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Ifc108981cb654c78e91f04e9055e8b4b6cb01052